requirements.txt
flask
flask-socketio
eventlet
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
